requirements.txt
numpy
matplotlib
tensorflow >=1.3; python_version == '3.5'
pillow
scipy
scikit-image